requirements.txt
streamlit
azure-ai-textanalytics
azure-core
azure-storage-blob
azure-cognitiveservices-speech
python-dotenv
openai
audio_recorder_streamlit